requirements.txt
langchain==0.1.17
langchain-community==0.0.36
langserve==0.0.41
langchain-openai==0.1.6
langchain-core==0.1.51

pydantic>=2.0
python-dotenv==1.0.1
streamlit==1.35.0
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
sse-starlette==1.6.5
